+++ packages/plugin-scroll/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @embedpdf/plugin-scroll
 
+## 2.6.2
+
+## 2.6.1
+
 ## 2.6.0
 
 ### Minor Changes