+++ viewers/react/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @embedpdf/react-pdf-viewer
 
+## 2.6.2
+
+## 2.6.1
+
 ## 2.6.0
 
 ## 2.5.0